requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.3.5
blinker==1.4
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
contextlib2==0.5.5
docopt==0.6.2
Flask==1.0.2
Flask-Dance==1.0.0
Flask-Login==0.4.1
Flask-OAuth==0.12
Flask-SQLAlchemy==2.3.2
future==0.16.0
google-auth==1.5.0
gunicorn==19.8.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
MarkupSafe==1.0
more-itertools==4.2.0
oauth2==1.9.0.post1
oauthlib==2.1.0
paho-mqtt==1.3.1
pbr==4.0.4
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.3
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycodestyle==2.4.0
PyJWT==1.6.4
pykwalify==1.6.1
pytest==3.6.2
python-box==3.2.0
python-dateutil==2.7.3
python-dotenv==0.8.2
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.10
SQLAlchemy-Utils==0.33.3
stevedore==1.28.0
tavern==0.12.4
urllib3==1.23
URLObject==2.4.3
Werkzeug==0.14.1
yapf==0.22.0
